requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-auth
requests
python-dotenv
